requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.24.0
openai==0.27.6
faiss-cpu==1.7.4
altair<5
